Added sound and music and fixed a quick bug
Added FMOD Unity integration, and set up the FMOD project in our Unity project. Now plays on start and is ready for integration with gameplay. Sounds do not play yet, but are also ready for integration with gameplay. This will not take long, once the code for the gamplay is written.

I also added a check to make sure an item isn't being held before changing back from mouseOverColor to originalColor.